retrain_package/training/requirements-train.txt
torch
torchvision
timm
albumentations
scikit-learn
mlflow
boto3
pydicom
opencv-python-headless
pillow
numpy
optuna